backend/requirements.txt
aiohttp==3.8.1
aionursery==0.3.0
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
black==22.3.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
DAWG-Python==0.7.2
docopt==0.6.2
factory-boy==3.2.1
fake-useragent==0.1.11
Faker==13.7.0
fastapi==0.75.1
fastapi-camelcase==1.0.5
flake8==4.0.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parse==1.19.0
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyee==8.2.2
pyflakes==2.4.0
pyhumps==3.5.3
pymongo==4.1.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.1.2
pytest-factoryboy==2.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
w3lib==1.22.0
websockets==10.3
yarl==1.7.2
zipp==3.8.0
